refactor: refatora criacao do servidor hapi.js
